Refactored tests a bit
